Added Missing connection Variables
Each Entry and Journal must have an Author and it's own ID.

The Author is passed through, and Journal will have it's own ID, and Entries will have their own, too.

When selecting a journal, it displays like JournalID entries. After selecting an Entry, when searching for it's history, it'll have the same EntryID as the previous version.